requirements.txt
Flask==3.1.0
gunicorn
PyJWT==2.8.0
python-dotenv==1.0.0
Pillow==10.2.0